working config.plist for sierra and highSierra
when using vga, just set inject intel to false, hd4600 was
not recognized by macos, but all things works with only 7MB
of video memory. when set inject intel to True, and ig-platform-id
set to proper string, and connect the screen to dvi or hdmi, hd4600
and video memory works as they were in a real mac :).